Working on PAUSEBOSS
we are trying to make it so that the boss stops for a certain amount of time before it continues after it uses its ability. true using the elapsed time thingy or something similr couritine dosent work its wac.

+++ Assets/Scripts/SpawnManager.cs
@@ -15,8 +15,8 @@ public class SpawnManager : MonoBehaviour
     public int waveCount = 1;
     public int bossCount;
     public int bossCounter;
-    public float totalEnimies;
-    public float enemyCount;
+    public int totalEnimies;
+    public int enemyCount;
 
 
     // Start is called before the first frame update
